requirements.txt
aiogram==3.7.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
alembic==1.13.1
g4f[all]
deep-translator==1.11.4
python-dotenv==1.0.1
aiohttp==3.9.5
